+++ dev/run-python-skinny-tests.sh
@@ -13,10 +13,9 @@ err=0
 trap 'err=1' ERR
 export MLFLOW_SKINNY='true'
 
-python -m pip uninstall sqlalchemy alembic sqlparse -y
 pytest --verbose tests/test_skinny.py
 python -m pip install sqlalchemy alembic sqlparse
-pytest --verbose tests/test_cli_run.py
+pytest --verbose tests/test_runs.py
 pytest --verbose tests/tracking/test_client.py
 pytest --verbose tests/tracking/test_tracking.py
 pytest --verbose tests/projects/test_projects.py
